# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Usta Cepte platform API — registration, support, ustas, requests, locations
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ustas:
    get:
      operationId: getUstas
      tags: [ustas]
      summary: List craftsmen
      parameters:
        - $ref: "#/components/parameters/DeviceIdOptional"
      responses:
        "200":
          description: List of craftsmen
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Usta"
  /ustas/{id}:
    get:
      operationId: getUstaById
      tags: [ustas]
      summary: Get craftsman details
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/DeviceIdOptional"
      responses:
        "200":
          description: Craftsman details with server-computed capability flags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usta"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ustas/{id}/status:
    put:
      operationId: updateUstaStatus
      tags: [ustas]
      summary: Update craftsman online status (owner only; unclaimed profiles require the claim code)
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/DeviceIdRequired"
        - name: x-claim-code
          in: header
          required: false
          description: Pre-authorized enrollment code, required to claim an unclaimed profile
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isOnline]
              properties:
                isOnline:
                  type: boolean
      responses:
        "200":
          description: Updated craftsman
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usta"
        "401":
          description: Missing device id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not the owner / invalid claim code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: number
    DeviceIdOptional:
      name: x-device-id
      in: header
      required: false
      description: Persistent per-device identifier; used to compute capability flags
      schema:
        type: string
    DeviceIdRequired:
      name: x-device-id
      in: header
      required: true
      description: Persistent per-device identifier; required for ownership-gated writes
      schema:
        type: string
  schemas:
...
    Usta:
      type: object
      required: [id, name, specialty, rating, priceMin, priceMax, isOnline, verified]
      properties:
        id:
          type: number
        name:
          type: string
        specialty:
          type: string
        categoryId:
          type: ["string", "null"]
        rating:
          type: number
        priceMin:
          type: number
        priceMax:
          type: number
        isOnline:
          type: boolean
        verified:
          type: boolean
        district:
          type: ["string", "null"]
        reviewCount:
          type: number
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        claimed:
          type: boolean
        canManage:
          type: boolean
        claimable:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string